+++ packages/advanced-logic/test/extensions/payment-network/ethereum/input-data.test.ts
@@ -124,7 +124,7 @@ describe('extensions/payment-network/ethereum/input-data', () => {
     describe('applyActionToExtension/unknown action', () => {
       it('cannot applyActionToExtensions of unknown action', () => {
         const unknownAction = Utils.deepCopy(DataEthAddPaymentAddress.actionAddPaymentAddress);
-        unknownAction.action = 'unknown action';
+        unknownAction.action = 'unknown action' as any;
         // 'must throw'
         expect(() => {
           ethereumInputData.applyActionToExtension(
@@ -139,7 +139,7 @@ describe('extensions/payment-network/ethereum/input-data', () => {
 
       it('cannot applyActionToExtensions of unknown id', () => {
         const unknownAction = Utils.deepCopy(DataEthAddPaymentAddress.actionAddPaymentAddress);
-        unknownAction.id = 'unknown id';
+        unknownAction.id = 'unknown id' as any;
         // 'must throw'
         expect(() => {
           ethereumInputData.applyActionToExtension(