+++ packages/ethereum-storage/src/ethereum-utils.ts
@@ -1,7 +1,16 @@
 import { StorageTypes } from '@requestnetwork/types';
 import * as config from './config';
 
-import * as BigNumber from 'bn.js';
+import { BigNumber } from 'ethers';
+
+const networks = {
+  [StorageTypes.EthereumNetwork.PRIVATE]: 'private',
+  [StorageTypes.EthereumNetwork.MAINNET]: 'mainnet',
+  [StorageTypes.EthereumNetwork.KOVAN]: 'kovan',
+  [StorageTypes.EthereumNetwork.RINKEBY]: 'rinkeby',
+  [StorageTypes.EthereumNetwork.SOKOL]: 'sokol',
+  [StorageTypes.EthereumNetwork.XDAI]: 'xdai',
+};
 
 /**
  * Collection of utils functions related to Ethereum
@@ -14,12 +23,15 @@ export default {
    * @return name of the network
    */
   getEthereumNetworkNameFromId(networkId: StorageTypes.EthereumNetwork): string {
-    return {
-      [StorageTypes.EthereumNetwork.PRIVATE as StorageTypes.EthereumNetwork]: 'private',
-      [StorageTypes.EthereumNetwork.MAINNET as StorageTypes.EthereumNetwork]: 'mainnet',
-      [StorageTypes.EthereumNetwork.KOVAN as StorageTypes.EthereumNetwork]: 'kovan',
-      [StorageTypes.EthereumNetwork.RINKEBY as StorageTypes.EthereumNetwork]: 'rinkeby',
-    }[networkId];
+    return networks[networkId];
+  },
+
+  getEthereumIdFromNetworkName(name: string): StorageTypes.EthereumNetwork | undefined {
+    const id = Object.entries(networks).find((entry) => entry[1] === name)?.[0];
+    if (!id) {
+      return undefined;
+    }
+    return Number(id) as StorageTypes.EthereumNetwork;
   },
 
   /**
@@ -32,8 +44,6 @@ export default {
    * @returns True if the gas price can be used
    */
   isGasPriceSafe(gasPrice: BigNumber): boolean {
-    return (
-      gasPrice.gt(new BigNumber(0)) && gasPrice.lt(new BigNumber(config.getSafeGasPriceLimit()))
-    );
+    return gasPrice.gt(0) && gasPrice.lt(config.getSafeGasPriceLimit());
   },
 };