+++ packages/ethereum-storage/src/gas-price-providers/ethgasstation-provider.ts
@@ -4,7 +4,7 @@ import { StorageTypes } from '@requestnetwork/types';
 import Utils from '@requestnetwork/utils';
 import fetch from 'node-fetch';
 
-import * as BigNumber from 'bn.js';
+import { BigNumber } from 'ethers';
 
 // Maximum number of api requests to retry when an error is encountered (ECONNRESET, EPIPE, ENOTFOUND)
 const ETHGASSTATION_REQUEST_MAX_RETRY = 3;
@@ -63,7 +63,7 @@ export default class EthGasStationProvider implements StorageTypes.IGasPriceProv
     }
 
     // Retrieve the gas price from the provided gas price type and the format of the API response
-    const apiGasPrice = new BigNumber(
+    const apiGasPrice = BigNumber.from(
       parseFloat(
         {
           [StorageTypes.GasPriceType.FAST]: apiResponse.fast,