+++ packages/ethereum-storage/test/gas-price-definer.test.ts
@@ -6,7 +6,7 @@ import EthereumUtils from '../src/ethereum-utils';
 import * as config from '../src/config';
 import GasPriceDefiner from '../src/gas-price-definer';
 
-import * as BigNumber from 'bn.js';
+import { BigNumber } from 'ethers';
 
 let gasPriceDefiner: GasPriceDefiner;
 
@@ -16,13 +16,35 @@ describe('GasPriceDefiner', () => {
   });
 
   describe('getGasPrice', () => {
-    it('returns default gas price from config if network is not mainnet', async () => {
+    it('returns default gas price from config if network has no provider', async () => {
+      gasPriceDefiner.gasPriceProviderList = {
+        [StorageTypes.EthereumNetwork.MAINNET]: [
+          { getGasPrice: () => Promise.resolve(BigNumber.from(1)) },
+        ],
+      };
       const gasPrice = await gasPriceDefiner.getGasPrice(
         StorageTypes.GasPriceType.STANDARD,
         EthereumUtils.getEthereumNetworkNameFromId(StorageTypes.EthereumNetwork.RINKEBY),
       );
 
-      expect(gasPrice).toBe(config.getDefaultEthereumGasPrice());
+      expect(gasPrice).toEqual(config.getDefaultEthereumGasPrice());
+    });
+
+    it('returns gas price from appropriate gas provider', async () => {
+      gasPriceDefiner.gasPriceProviderList = {
+        [StorageTypes.EthereumNetwork.MAINNET]: [
+          { getGasPrice: () => Promise.resolve(BigNumber.from(1)) },
+        ],
+        [StorageTypes.EthereumNetwork.XDAI]: [
+          { getGasPrice: () => Promise.resolve(BigNumber.from(2)) },
+        ],
+      };
+      const gasPrice = await gasPriceDefiner.getGasPrice(
+        StorageTypes.GasPriceType.STANDARD,
+        EthereumUtils.getEthereumNetworkNameFromId(StorageTypes.EthereumNetwork.XDAI),
+      );
+
+      expect(gasPrice).toEqual(BigNumber.from(2));
     });
 
     it('returns default gas price from config if no provider is available', async () => {
@@ -34,88 +56,80 @@ describe('GasPriceDefiner', () => {
         EthereumUtils.getEthereumNetworkNameFromId(StorageTypes.EthereumNetwork.MAINNET),
       );
 
-      expect(gasPrice).toBe(config.getDefaultEthereumGasPrice());
+      expect(gasPrice).toEqual(config.getDefaultEthereumGasPrice());
     });
 
     it('returns the max of values returned by providers', async () => {
-      gasPriceDefiner.gasPriceProviderList = [
-        {
-          getGasPrice: async (_type: StorageTypes.GasPriceType): Promise<BigNumber> =>
-            new BigNumber(100),
-          providerUrl: '',
-        },
-        {
-          getGasPrice: async (_type: StorageTypes.GasPriceType): Promise<BigNumber> =>
-            new BigNumber(200),
-          providerUrl: '',
-        },
-        {
-          getGasPrice: async (_type: StorageTypes.GasPriceType): Promise<BigNumber> =>
-            new BigNumber(300),
-          providerUrl: '',
-        },
-        {
-          getGasPrice: async (_type: StorageTypes.GasPriceType): Promise<BigNumber> =>
-            new BigNumber(40),
-          providerUrl: '',
-        },
-        {
-          getGasPrice: async (_type: StorageTypes.GasPriceType): Promise<BigNumber> =>
-            new BigNumber(300),
-          providerUrl: '',
-        },
-      ];
+      gasPriceDefiner.gasPriceProviderList = {
+        [StorageTypes.EthereumNetwork.MAINNET]: [
+          { getGasPrice: () => Promise.resolve(BigNumber.from(100)) },
+          { getGasPrice: () => Promise.resolve(BigNumber.from(200)) },
+          { getGasPrice: () => Promise.resolve(BigNumber.from(300)) },
+          { getGasPrice: () => Promise.resolve(BigNumber.from(40)) },
+          { getGasPrice: () => Promise.resolve(BigNumber.from(300)) },
+        ],
+      };
 
       const gasPrice = await gasPriceDefiner.getGasPrice(
         StorageTypes.GasPriceType.STANDARD,
         EthereumUtils.getEthereumNetworkNameFromId(StorageTypes.EthereumNetwork.MAINNET),
       );
 
-      expect(gasPrice).toBe('300');
+      expect(gasPrice).toEqual(BigNumber.from(300));
     });
   });
 
   describe('pollProviders', () => {
     it('returns an array containing value from each provider', async () => {
-      gasPriceDefiner.gasPriceProviderList = [
-        {
-          getGasPrice: async (_type: StorageTypes.GasPriceType): Promise<BigNumber> =>
-            new BigNumber(100),
-          providerUrl: '',
-        },
-        {
-          getGasPrice: async (_type: StorageTypes.GasPriceType): Promise<BigNumber> =>
-            new BigNumber(500),
-          providerUrl: '',
-        },
-        {
-          getGasPrice: async (_type: StorageTypes.GasPriceType): Promise<BigNumber> =>
-            new BigNumber(200),
-          providerUrl: '',
-        },
-        {
-          getGasPrice: async (_type: StorageTypes.GasPriceType): Promise<BigNumber> =>
-            new BigNumber(300000),
-          providerUrl: '',
-        },
-      ];
+      gasPriceDefiner.gasPriceProviderList = {
+        [StorageTypes.EthereumNetwork.MAINNET]: [
+          { getGasPrice: () => Promise.resolve(BigNumber.from(100)) },
+          { getGasPrice: () => Promise.resolve(BigNumber.from(500)) },
+          { getGasPrice: () => Promise.resolve(BigNumber.from(200)) },
+          { getGasPrice: () => Promise.resolve(BigNumber.from(300000)) },
+        ],
+      };
 
       await expect(
-        gasPriceDefiner.pollProviders(StorageTypes.GasPriceType.STANDARD),
+        gasPriceDefiner.pollProviders(
+          StorageTypes.GasPriceType.STANDARD,
+          StorageTypes.EthereumNetwork.MAINNET,
+        ),
       ).resolves.toEqual([
-        new BigNumber(100),
-        new BigNumber(500),
-        new BigNumber(200),
-        new BigNumber(300000),
+        BigNumber.from(100),
+        BigNumber.from(500),
+        BigNumber.from(200),
+        BigNumber.from(300000),
       ]);
     });
 
     it('returns empty array if there is no provider', async () => {
-      gasPriceDefiner.gasPriceProviderList = [];
+      gasPriceDefiner.gasPriceProviderList = { [StorageTypes.EthereumNetwork.MAINNET]: [] };
 
       await expect(
-        gasPriceDefiner.pollProviders(StorageTypes.GasPriceType.STANDARD),
+        gasPriceDefiner.pollProviders(
+          StorageTypes.GasPriceType.STANDARD,
+          StorageTypes.EthereumNetwork.MAINNET,
+        ),
       ).resolves.toHaveLength(0);
+    });
+
+    it('handles failures and invalid returns', async () => {
+      gasPriceDefiner.gasPriceProviderList = {
+        [StorageTypes.EthereumNetwork.MAINNET]: [
+          { getGasPrice: () => Promise.resolve(BigNumber.from(100)) },
+          { getGasPrice: () => Promise.reject(new Error('oops!')) },
+          { getGasPrice: () => Promise.resolve(BigNumber.from(500)) },
+          { getGasPrice: () => Promise.resolve(null) },
+        ],
+      };
+
+      await expect(
+        gasPriceDefiner.pollProviders(
+          StorageTypes.GasPriceType.STANDARD,
+          StorageTypes.EthereumNetwork.MAINNET,
+        ),
+      ).resolves.toEqual([BigNumber.from(100), BigNumber.from(500)]);
     });
   });
 });