+++ packages/currency/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.5.0 (2021-03-25)
+
+
+### Bug Fixes
+
+* ERC20 currency case ([#428](https://github.com/RequestNetwork/requestNetwork/issues/428)) ([9f4f131](https://github.com/RequestNetwork/requestNetwork/commit/9f4f13114d32ad30cedee2187052a41f16a8985f))
+
+
+### Features
+
+* add mph and ausdc tokens to supported tokens ([#423](https://github.com/RequestNetwork/requestNetwork/issues/423)) ([6f8afec](https://github.com/RequestNetwork/requestNetwork/commit/6f8afec5bc5fbbd8b33280808bff47d07e5377f0))
+* add OCEAN and SAND ([#451](https://github.com/RequestNetwork/requestNetwork/issues/451)) ([055530c](https://github.com/RequestNetwork/requestNetwork/commit/055530c5326489e633034f3f2b7940bb39a39aaa))
+* payment network any to erc20 smartcontracts ([#408](https://github.com/RequestNetwork/requestNetwork/issues/408)) ([a2f30a8](https://github.com/RequestNetwork/requestNetwork/commit/a2f30a84689eaea0994e72944c417718c7aad20e))
+
+
+
+
+
+# 0.4.0 (2021-03-15)
+
+
+### Bug Fixes
+
+* ERC20 currency case ([#428](https://github.com/RequestNetwork/requestNetwork/issues/428)) ([9f4f131](https://github.com/RequestNetwork/requestNetwork/commit/9f4f13114d32ad30cedee2187052a41f16a8985f))
+
+
+### Features
+
+* add mph and ausdc tokens to supported tokens ([#423](https://github.com/RequestNetwork/requestNetwork/issues/423)) ([6f8afec](https://github.com/RequestNetwork/requestNetwork/commit/6f8afec5bc5fbbd8b33280808bff47d07e5377f0))
+* payment network any to erc20 smartcontracts ([#408](https://github.com/RequestNetwork/requestNetwork/issues/408)) ([a2f30a8](https://github.com/RequestNetwork/requestNetwork/commit/a2f30a84689eaea0994e72944c417718c7aad20e))
+
+
+
+
+
 # 0.3.0 (2021-03-03)